requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.104.1
fonttools==4.57.0
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.3
opencv-python==4.12.0.88
opencv-python-headless==4.8.1.78
packaging==25.0
pandas==2.0.3
Pillow==9.5.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
requests==2.31.0
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.176
ultralytics-thop==2.0.15
urllib3==2.2.3
uvicorn==0.24.0
watchfiles==0.24.0
websockets==13.1
zipp==3.20.2